requirements.txt
yfinance>=0.2.31
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
optuna>=3.4.0
lxml>=4.9.0
alpaca-py>=0.21.0
